apps/ml-service/ml_requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
transformers==4.38.2
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.1+cpu
torchvision==0.17.1+cpu
pillow>=10.1.0
python-multipart>=0.0.6
numpy<2
